refactor(parser): extract arg-metadata into types/arg_metadata
Plan 2, step 5 of N. Move `extract_arg_metadata` + the help-section
helpers (`resolve_help_section`, `is_arg_section_call_expr`,
`parse_inline_arg_section`) into
`crates/toolr-core/src/parser/types/arg_metadata.rs` (148 lines).

The new module reaches `is_toolr_arg_call` and `literal_*` via
`super::` — both were promoted to `pub(super)` in earlier steps for
exactly this case.

`types/mod.rs` shrinks 920 → 788 lines. The `HelpSection` import
moved with the consumer; the previously-needed
`use self::literals::{literal_str, literal_str_list, literal_u32};`
in mod.rs is gone now that arg_metadata.rs is the sole caller.

Public surface preserved via `pub use arg_metadata::extract_arg_metadata`.

Tests: 24 / 24 parser::types::tests still pass — including the six
extract_arg_metadata_* tests in the existing test module, which
exercise the moved function via the re-export.
